refactor: Extract useDualAxisMargins
