+++ app/charts/combo/combo-line-dual-state.tsx
@@ -9,15 +9,11 @@ import {
 } from "@/charts/combo/combo-line-dual-state-props";
 import {
   adjustScales,
-  getMargins,
   useCommonComboState,
+  useDualAxisMargins,
   useYScales,
 } from "@/charts/combo/combo-state";
-import {
-  SINGLE_LINE_AXIS_LABEL_HEIGHT,
-  TITLE_V_PADDING,
-  useAxisTitleSize,
-} from "@/charts/combo/shared";
+import { TITLE_V_PADDING } from "@/charts/combo/shared";
 import { TICK_PADDING } from "@/charts/shared/axis-height-linear";
 import {
   useChartBounds,
@@ -134,31 +130,18 @@ const useComboLineDualState = (
         TICK_PADDING
     )
   );
-  const axisTitleWidth = width * 0.5 - TICK_PADDING;
-  const leftAxisTitleSize = useAxisTitleSize(variables.y.left.label, {
-    width: axisTitleWidth,
-  });
-  const rightAxisTitleSize = useAxisTitleSize(variables.y.right.label, {
-    width: width * 0.5 - TICK_PADDING,
-  });
-  const topAdd =
-    Math.max(leftAxisTitleSize.height, rightAxisTitleSize.height) +
-    SINGLE_LINE_AXIS_LABEL_HEIGHT;
-
-  const right = Math.max(maxRightTickWidth, 40);
-
-  const margins = getMargins({
+  const margins = useDualAxisMargins({
+    width,
     left,
-    right,
-    bottom: bottom - SINGLE_LINE_AXIS_LABEL_HEIGHT,
-    top: topAdd,
+    bottom,
+    maxRightTickWidth,
+    leftAxisTitle: variables.y.left.label,
+    rightAxisTitle: variables.y.right.label,
   });
-
   const bounds = useChartBounds(width, margins, height, {
     leftLabel: variables.y.left.label,
     rightLabel: variables.y.right.label,
   });
-
   const { chartWidth, chartHeight } = bounds;
   const xScales = [xScale, xScaleTimeRange];
   const yScales = [yScale, yScaleLeft, yScaleRight];
